367: Reorganize models moving the commenting-requests to their own sub-directory
