+++ src/models/requests/commenting/index.ts
@@ -1,0 +1,3 @@
+export type {CommentFilterRequest} from './CommentFilterRequest';
+export type {CommentRequest} from './CommentRequest';
+export type {ReportRequest} from './ReportRequest';
